# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mail Monitor API",
        "description": "API for Exim/Dovecot monitoring and metrics.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current host"
        }
    ],
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "7ca6fc7989202054d29707fef91a2d4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthRegisterResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/v1/auth/token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login and issue token",
                "operationId": "a80f60d957dabb031fa22b5915096212",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthLoginResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke current token",
                "operationId": "3a27a9e02ceb521eb3f98c8d2394aef9",
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/mailboxes": {
            "get": {
                "tags": [
                    "Mailboxes"
                ],
                "summary": "List mailboxes",
                "operationId": "148f144addd98cdce212d1976ef62bb3",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 200
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedMailboxes"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mailboxes"
                ],
                "summary": "Create mailbox",
                "operationId": "cb674864ec34a43bf00d39292a7292b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MailboxStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mailbox"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/mailboxes/{mailbox}": {
            "patch": {
                "tags": [
                    "Mailboxes"
                ],
                "summary": "Update mailbox",
                "operationId": "2310117a8c531dd308c5edc038f70b8d",
                "parameters": [
                    {
                        "name": "mailbox",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MailboxUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mailbox"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CpanelAccount": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "whm_account": {
                        "type": "string",
                        "example": "rmlad479"
                    },
                    "domain": {
                        "type": "string",
                        "example": "rmladv.com.br"
                    },
                    "cpanel_host": {
                        "type": "string",
                        "example": "https://server:2087"
                    }
                },
                "type": "object"
            },
            "AuthRegisterResponse": {
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "cpanel_account": {
                        "$ref": "#/components/schemas/CpanelAccount"
                    }
                },
                "type": "object"
            },
            "AuthLoginResponse": {
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "whm_account",
                    "domain",
                    "cpanel_host",
                    "cpanel_api_token"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password"
                    },
                    "whm_account": {
                        "type": "string"
                    },
                    "domain": {
                        "type": "string"
                    },
                    "cpanel_host": {
                        "type": "string"
                    },
                    "cpanel_api_token": {
                        "type": "string"
                    },
                    "cpanel_username": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
...
            "Mailbox": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "whm_account": {
                        "type": "string"
                    },
                    "domain": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PaginatedMailboxes": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Mailbox"
                        }
                    },
                    "meta": {
                        "type": "object"
                    },
                    "links": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "MailboxStoreRequest": {
                "required": [
                    "whm_account",
                    "domain",
                    "email"
                ],
                "properties": {
                    "whm_account": {
                        "type": "string"
                    },
                    "domain": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "MailboxUpdateRequest": {
                "properties": {
                    "whm_account": {
                        "type": "string"
                    },
                    "domain": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Informe o token Bearer retornado pela API."
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}